[sdk] Reduce circular deps
