+++ packages/sdk/src/evm/core/classes/index.ts
@@ -11,13 +11,7 @@ export * from "./erc-20";
 export * from "./erc-20-history";
 export * from "./erc-20-standard";
 export * from "./erc-721";
-export * from "./erc-721-lazymintable";
-export * from "./erc-721-claimable-with-conditions";
 export * from "./erc-721-with-quantity-signature-mintable";
-export * from "./erc-721-supply";
-export * from "./erc-721-enumerable";
-export * from "./erc-721-mintable";
-export * from "./erc-721-batch-mintable";
 export * from "./erc-721-burnable";
 export * from "./erc-721-standard";
 export * from "./erc-1155";
